# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.CardRequest:
    properties:
      balance:
        type: number
      card_number:
        type: string
      id:
        type: integer
      is_blocked:
        type: integer
      key_id:
        type: integer
      owner_name:
        type: string
    type: object
  handlers.KeyRequest:
    properties:
      id:
        type: integer
      value:
        type: string
    type: object
...
  repository.Card:
    properties:
      balance:
        type: number
      cardNumber:
        type: string
      id:
        type: integer
      isBlocked:
        type: integer
      keyID:
        type: integer
      ownerName:
        type: string
    type: object
  repository.Key:
    properties:
      id:
        type: integer
      value:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Lab2 REST API
  version: "1.0"
paths:
...
  /cards/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repository.Card'
            type: array
        "405":
          description: Method not allowed
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List cards
      tags:
      - cards
  /cards/balance:
    put:
      consumes:
      - application/json
      description: Admin only. Sets balance to the given value.
      parameters:
      - description: Bearer admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: id and balance
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            type: string
        "401":
          description: Missing or invalid token
          schema:
            type: string
        "403":
          description: Not an admin
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update card balance only
      tags:
      - cards
  /cards/create:
    post:
      consumes:
      - application/json
      description: Admin only.
      parameters:
      - description: Bearer admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: card_number, balance, is_blocked, owner_name, optional key_id
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            type: string
        "401":
          description: Missing or invalid token
          schema:
            type: string
        "403":
          description: Not an admin
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create card
      tags:
      - cards
  /cards/delete:
    delete:
      description: Admin only.
      parameters:
      - description: Bearer admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: Card ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid id
          schema:
            type: string
        "401":
          description: Missing or invalid token
          schema:
            type: string
        "403":
          description: Not an admin
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete card
      tags:
      - cards
  /cards/get:
    get:
      parameters:
      - description: Card ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.Card'
        "400":
          description: Invalid id or not found
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      summary: Get card by ID
      tags:
      - cards
  /cards/update:
    put:
      consumes:
      - application/json
      description: Admin only. Updates balance, block flag, owner, and key link.
      parameters:
      - description: Bearer admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: id and fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            type: string
        "401":
          description: Missing or invalid token
          schema:
            type: string
        "403":
          description: Not an admin
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update card
      tags:
      - cards
  /keys/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repository.Key'
            type: array
        "405":
          description: Method not allowed
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List keys
      tags:
      - keys
  /keys/create:
    post:
      consumes:
      - application/json
      description: Admin only.
      parameters:
      - description: Bearer admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: value
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.KeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            type: string
        "401":
          description: Missing or invalid token
          schema:
            type: string
        "403":
          description: Not an admin
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create key
      tags:
      - keys
  /keys/delete:
    delete:
      description: Admin only.
      parameters:
      - description: Bearer admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: Key ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid id
          schema:
            type: string
        "401":
          description: Missing or invalid token
          schema:
            type: string
        "403":
          description: Not an admin
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete key
      tags:
      - keys
  /keys/get:
    get:
      parameters:
      - description: Key ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.Key'
        "400":
          description: Invalid id or not found
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      summary: Get key by ID
      tags:
      - keys
  /keys/update:
    put:
      consumes:
      - application/json
      description: Admin only.
      parameters:
      - description: Bearer admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: id, value
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.KeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            type: string
        "401":
          description: Missing or invalid token
          schema:
            type: string
        "403":
          description: Not an admin
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update key
      tags:
      - keys
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"